chore: migrate delimiters to AST
